Fix line spacing in lectures
Previously, the line height between two lines of the same list item and two different list items would be the same, in all sections except the topics list. Now there is proper and uniform spacing (with more space between list items than between lines in the same list item) for topics, references, and slides and notes. This is most noticeable on narrower screens.